Split interpreter / synthesiser
