+++ src/RamExecutor.h
@@ -36,6 +36,59 @@ namespace souffle {
 class RamStatement;
 class RamInsert;
 
+class EvalContext {
+    std::vector<const RamDomain*> data;
+    std::vector<RamDomain>* returnValues = nullptr;
+    std::vector<bool>* returnErrors = nullptr;
+    const std::vector<RamDomain>* args = nullptr;
+
+public:
+    EvalContext(size_t size = 0) : data(size) {}
+
+    const RamDomain*& operator[](size_t index) {
+        return data[index];
+    }
+
+    const RamDomain* const& operator[](size_t index) const {
+        return data[index];
+    }
+
+    std::vector<RamDomain>& getReturnValues() const {
+        return *returnValues;
+    }
+
+    void setReturnValues(std::vector<RamDomain>& retVals) {
+        returnValues = &retVals;
+    }
+
+    void addReturnValue(RamDomain val, bool err = false) {
+        assert(returnValues != nullptr && returnErrors != nullptr);
+        returnValues->push_back(val);
+        returnErrors->push_back(err);
+    }
+
+    std::vector<bool>& getReturnErrors() const {
+        return *returnErrors;
+    }
+
+    void setReturnErrors(std::vector<bool>& retErrs) {
+        returnErrors = &retErrs;
+    }
+
+    const std::vector<RamDomain>& getArguments() const {
+        return *args;
+    }
+
+    void setArguments(const std::vector<RamDomain>& a) {
+        args = &a;
+    }
+
+    RamDomain getArgument(size_t i) const {
+        assert(args != nullptr && i < args->size() && "argument out of range");
+        return (*args)[i];
+    }
+};
+
 /**
  * An abstract base class for entities capable of processing a RAM program.
  */
@@ -88,9 +141,9 @@ public:
     /**
      * Runs a subroutine of a RamProgram
      */
-    void executeSubroutine(RamEnvironment& env, const RamStatement& stmt,
+    virtual void executeSubroutine(RamEnvironment& env, const RamStatement& stmt,
             const std::vector<RamDomain>& arguments, std::vector<RamDomain>& returnValues,
-            std::vector<bool>& returnErrors);
+            std::vector<bool>& returnErrors) = 0; 
 
     /**
      * Runs the given statement on the given environment.
@@ -188,6 +241,13 @@ public:
      * on the given environment.
      */
     void applyOn(const RamProgram& prog, RamEnvironment& env, RamData* data) const override;
+
+    /**
+     * Runs a subroutine of a RamProgram
+     */
+    virtual void executeSubroutine(RamEnvironment& env, const RamStatement& stmt,
+            const std::vector<RamDomain>& arguments, std::vector<RamDomain>& returnValues,
+            std::vector<bool>& returnErrors) override;
 };
 
 /**
@@ -258,6 +318,11 @@ public:
      */
     // void compileSubroutine(std::string name, const RamStatement& stmt, IndexMap& indices, std::ostream& os)
     // const;
+    virtual void executeSubroutine(RamEnvironment& env, const RamStatement& stmt,
+            const std::vector<RamDomain>& arguments, std::vector<RamDomain>& returnValues,
+            std::vector<bool>& returnErrors) override {
+        // nothing to do here
+    }
 };
 
 /**
